RAG_V2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain
langchain-community
langchain-text-splitters==0.0.1
groq==0.4.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2